Added Scalability and more parameters
Added the ability to select a number of deployments. It uses a for loop to create all three machines, then write over the files to increment a number at the end of all necessary resource names, then loop. At the end it will clean up the files by setting all of those names back to 1. Also added elapsed times